Fix compiler warning
Signed\unsigned mismatch.

+++ src/ref_soft/r_rast.c
@@ -598,7 +598,7 @@ void R_RenderFace (msurface_t *fa, int clipflags)
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED)
 				{
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount)
+						( unsigned int )r_framecount)
 					{
 						r_lastvertvalid = false;
 						continue;
@@ -642,7 +642,7 @@ void R_RenderFace (msurface_t *fa, int clipflags)
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED)
 				{
 					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) ==
-						r_framecount)
+						( unsigned int )r_framecount)
 					{
 						r_lastvertvalid = false;
 						continue;
